chore: consolidate auth context
